// Frontend/Web/src/gen/files_api_pb.ts
export enum UploadFileType {
  /**
   * @generated from enum value: UPLOAD_FILE_TYPE_UNKNOWN = 0;
   */
  UPLOAD_FILE_TYPE_UNKNOWN = 0,

  /**
   * Аватарка пользователя
   *
   * @generated from enum value: USER_AVATAR = 1;
   */
  USER_AVATAR = 1,

  /**
   * Вложение сообщения - изображение
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_IMAGE = 2;
   */
  MESSAGE_ATTACHMENT_IMAGE = 2,

  /**
   * Вложение сообщения - видео
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_VIDEO = 3;
   */
  MESSAGE_ATTACHMENT_VIDEO = 3,

  /**
   * Вложение сообщения - гифка
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_GIF = 4;
   */
  MESSAGE_ATTACHMENT_GIF = 4,

  /**
   * Вложение сообщения - документ
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_DOCUMENT = 5;
   */
  MESSAGE_ATTACHMENT_DOCUMENT = 5,

  /**
   * Обложка группового чата
   *
   * @generated from enum value: CHAT_PICTURE = 6;
   */
  CHAT_PICTURE = 6,

  /**
   * Вложение сообщения - аудио
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_AUDIO = 7;
   */
  MESSAGE_ATTACHMENT_AUDIO = 7,

  /**
   * Вложение сообщения - голосовое сообщение
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_VOICE = 8;
   */
  MESSAGE_ATTACHMENT_VOICE = 8,

  /**
   * Вложение сообщения - стикер (WebP)
   *
   * @generated from enum value: MESSAGE_ATTACHMENT_STICKER = 9;
   */
  MESSAGE_ATTACHMENT_STICKER = 9,

  /**
   * Постер профиля пользователя (горизонтальный баннер)
   *
   * @generated from enum value: USER_PROFILE_POSTER = 10;
   */
  USER_PROFILE_POSTER = 10,
}